新增了UI！现在我们真正能看到英雄的状态了！同时将一些有关时间的变量、函数统一修正为float。

+++ Classes/Actor/Assassin.cpp
@@ -6,7 +6,7 @@
 #include "Assassin.h"
 #include "Item/Shotgun.h"
 #include "Item/Sword.h"
-#include "Buff.h"
+#include "Component/Buff.h"
 
 bool Assassin::init()
 {
@@ -30,7 +30,7 @@ bool Assassin::init()
 	}
 }
 
-double Assassin::skill()
+float Assassin::skill()
 {
 	if (m_ifMoved)
 	{
